llm-python/pyproject.toml
[project]
name = "o3-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.0.0",
    "huggingface-hub>=0.33.4",
    "matplotlib>=3.7.0",
    "numpy>=2.3.1",
    "openai>=1.93.0",
    "pandas>=2.3.1",
    "pillow>=10.0.0",
    "pip>=25.1.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "tensorboard>=2.20.0",
    "tiktoken>=0.9.0",
]
